refactor: modularize auth tier and serverKeys code
Split large auth files into modular structure:

src/auth/tier/
├── types.ts      - Schemas and type definitions
├── cache.ts      - TTL cache management
├── remote.ts     - Remote config fetching from /relay/tier-config
├── validation.ts - Model/provider validation functions
└── index.ts      - Public exports

src/auth/serverKeys/
├── types.ts      - Type definitions
├── cache.ts      - Provider and access cache management
├── settings.ts   - User preference (included vs personal keys)
├── providers.ts  - Fetching enabled providers
├── access.ts     - Access determination functions
├── routing.ts    - Relay URL generation
└── index.ts      - Public exports

Benefits:
- Better separation of concerns
- Easier to test individual modules
- More maintainable code structure
- Backward compatible via re-exports from old files
